feat(migrations): native SQL migration engine (#7–#22)
Implements CREATE MIGRATION, APPLY MIGRATION, ROLLBACK MIGRATION, and
EXPLAIN MIGRATION as first-class SQL commands — no external tools needed.

Core modules:
- migration_collections: red_migrations + red_migration_deps system collections
- migration_graph: DAG topological sort (Kahn) + cycle detection (DFS)
- migration_inference: static SQL body analysis for auto-inferred deps
- impl_migrations: full execution engine (create, apply, rollback, explain)
- parser/migration: SQL parser for all four migration commands

Key features:
- DEPENDS ON for explicit dependency edges; static analysis fills gaps
- APPLY MIGRATION * applies all pending in topological order
- BATCH N ROWS for data migrations with rows_processed checkpoint (resume on crash)
- NO ROLLBACK flag for irreversible migrations
- Multi-statement DDL bodies (semicolon-separated)
- FOR TENANT <id> scopes apply to a specific tenant; FOR TENANT * fans out
- VCS commit recorded per applied migration (vcs_commit_hash field)
- Permission gates: CREATE requires Write, APPLY/ROLLBACK require Admin

Also fixes stale gRPC default port in test assertions (50051 → 5055).

+++ src/storage/timeseries/compression.rs
@@ -286,7 +286,7 @@ pub fn select_int_codec(values: &[i64]) -> TsIntCodec {
     let bits = if range == 0 {
         0
     } else {
-        128 - range.leading_zeros() as u32
+        128 - range.leading_zeros()
     };
     if bits <= 20 {
         return TsIntCodec::T64;
